+++ movie-inventory-service/src/main/java/com/example/movieinventoryservice/modules/movies/service/serviceImpl/GenreServiceImpl.java
@@ -3,6 +3,8 @@ package com.example.movieinventoryservice.modules.movies.service.serviceImpl;
 import java.util.List;
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
@@ -14,6 +16,7 @@ import com.example.movieinventoryservice.exception.RecordNotAddedException;
 import com.example.movieinventoryservice.exception.RecordNotDeletedException;
 import com.example.movieinventoryservice.exception.RecordNotFoundException;
 import com.example.movieinventoryservice.exception.RecordNotUpdatedException;
+import com.example.movieinventoryservice.exception.ServiceException;
 import com.example.movieinventoryservice.modules.movies.repository.GenreRepository;
 import com.example.movieinventoryservice.modules.movies.service.GenreService;
 
@@ -24,51 +27,62 @@ public class GenreServiceImpl implements GenreService {
 	@Autowired
 	private GenreRepository genreRepository;
 	
-	private String message = "";
+	private Logger logger = LoggerFactory.getLogger(GenreServiceImpl.class);
 
 	
 
 	@Override
-	public String deleteGenre(int genreId) throws RecordNotDeletedException {
+	public void deleteGenre(int genreId) throws RecordNotDeletedException,ServiceException {
 		try {
+			if(getGenreById(genreId)!=null) {
 			genreRepository.deleteById(genreId);
-			message = "Successfully deleted Genre of id - " + genreId;
-			return message;
-		} catch (DataAccessException ex) {
-			throw new RecordNotDeletedException("Failed to Delete", ex.getCause());
+			}
+			}catch (RecordNotFoundException e) {
+				throw new RecordNotDeletedException("genred id - "+genreId+" not found");
+			}
+		catch (DataAccessException ex) {
+			throw new ServiceException("Failed to Delete id-"+genreId, ex.getCause());
 		}
 	}
 
 	@Override
-	public String updateGenre(Genre genre) throws RecordNotUpdatedException {
+	public Genre updateGenre(Genre genre) throws RecordNotUpdatedException,ServiceException {
 		try {
-
-			genreRepository.save(genre);
-			message = "Successfully updated genre of name - " + genre.getName();
-			return message;
+			if(getGenreById(genre.getId())!=null) {
+			return genreRepository.save(genre);
+			}
+		}catch(RecordNotFoundException ex) {
+			throw new RecordNotUpdatedException("genred id - "+genre.getId()+" not found");
 		}
 
 		catch (DataAccessException ex) {
+			throw new ServiceException("Failed to Update", ex.getCause());
+		}
+		catch(Exception ex) {
 			throw new RecordNotUpdatedException("Failed to Update", ex.getCause());
 		}
+		return genre;
 	}
 
 	@Override
-	public Genre getGenreById(int genreId) throws RecordNotFoundException {
+	public Genre getGenreById(int genreId) throws RecordNotFoundException ,ServiceException{
 		try {
+			logger.info("Entered into Genre Service - getByid "+genreId);
+			logger.info(genreRepository.findAll().toString());
 			Optional<Genre> genre = genreRepository.findById(genreId);
 			if (genre.isPresent()) {
+				logger.info(genre.get().toString());
 				return genre.get();
 			} else {
-				throw new RecordNotFoundException("No Record to Fetch");
+				throw new RecordNotFoundException("Genre id -"+genreId+"is not present");
 			}
 		} catch (DataAccessException ex) {
-			throw new RecordNotFoundException("Failed to Fetch", ex.getCause());
+			throw new ServiceException("Failed to Fetch", ex.getCause());
 		}
 	}
 
 	@Override
-	public List<Genre> getAllGenres() throws EmptyListException {
+	public List<Genre> getAllGenres() throws EmptyListException,ServiceException {
 		try {
 			List<Genre> genres = genreRepository.findAll();
 			if (genres.size()>0) {
@@ -77,16 +91,18 @@ public class GenreServiceImpl implements GenreService {
 				throw new EmptyListException("No Record to Fetch");
 			}
 		} catch (DataAccessException ex) {
-			throw new EmptyListException("Failed to Fetch", ex.getCause());
+			throw new ServiceException("Failed to Fetch", ex.getCause());
 		}
-	}
+	}	
 
 	@Override
-	public Genre addGenre(Genre genre) throws RecordNotAddedException {
+	public Genre addGenre(Genre genre) throws RecordNotAddedException, ServiceException {
 		try {
 			return genreRepository.save(genre);
 		} catch (DataAccessException ex) {
-			throw new RecordNotAddedException("Failed To Add Genre", ex.getCause());
+			throw new ServiceException("Failed To Add Genre", ex.getCause());
+		}catch(Exception ex) {
+			throw new RecordNotAddedException("Invalid Data");
 		}
 	}
