+++ movie-inventory-service/src/main/java/com/example/movieinventoryservice/modules/movies/service/serviceImpl/MovieServiceImpl.java
@@ -17,6 +17,7 @@ import com.example.movieinventoryservice.exception.RecordNotAddedException;
 import com.example.movieinventoryservice.exception.RecordNotDeletedException;
 import com.example.movieinventoryservice.exception.RecordNotFoundException;
 import com.example.movieinventoryservice.exception.RecordNotUpdatedException;
+import com.example.movieinventoryservice.exception.ServiceException;
 import com.example.movieinventoryservice.modules.movies.repository.CastRepository;
 import com.example.movieinventoryservice.modules.movies.repository.GenreRepository;
 import com.example.movieinventoryservice.modules.movies.repository.MovieRepository;
@@ -25,10 +26,8 @@ import com.example.movieinventoryservice.modules.movies.service.MovieService;
 @Service
 
 public class MovieServiceImpl implements MovieService {
-	
-	private Logger logger = LoggerFactory.getLogger(MovieServiceImpl.class);
 
-	private String message = "";
+	private Logger logger = LoggerFactory.getLogger(MovieServiceImpl.class);
 
 	@Autowired
 	private MovieRepository movieRepository;
@@ -41,98 +40,104 @@ public class MovieServiceImpl implements MovieService {
 
 	@Override
 	@Transactional
-	public Movie addMovie(Movie movie) throws RecordNotAddedException {
+	public Movie addMovie(Movie movie) throws RecordNotAddedException, ServiceException {
 
 		try {
 
-			
-			movie.getGenre().forEach(val->{
-				Optional<Integer>genreId = genreRepository.getIdByName(val.getName());
-				if(genreId.isPresent()) {
+			movie.getGenre().forEach(val -> {
+				Optional<Integer> genreId = genreRepository.getIdByName(val.getName());
+				if (genreId.isPresent()) {
 					val.setId(genreId.get());
 				}
 			});
 			movie.getGenre().forEach(v -> genreRepository.save(v));
-			
+
 			logger.info(movie.getGenre().toString());
 
 			List<Cast> cast = movie.getCast();
 			cast.forEach(v -> v.setId(castRepository.save(v).getId()));
-			
+
 			logger.info(movie.getCast().toString());
 
 			return movieRepository.save(movie);
 		} catch (DataAccessException ex) {
 			System.out.println(ex.getLocalizedMessage());
-			throw new RecordNotAddedException("Failed To Add Movie", ex.getCause());
+			throw new ServiceException("Failed To Add Movie", ex.getCause());
+		} catch (Exception ex) {
+			throw new RecordNotAddedException("Invalid Data");
 		}
 
 	}
 
 	@Override
-	public String deleteMovie(int movieId) throws RecordNotDeletedException {
+	public void deleteMovie(int movieId) throws RecordNotDeletedException, ServiceException {
 		try {
-			movieRepository.deleteById(movieId);
-			message = "Successfully deleted movie of id - " + movieId;
-			return message;
+			if (getMovieById(movieId) != null) {
+				movieRepository.deleteById(movieId);
+			}
+		} catch (RecordNotFoundException e) {
+			throw new RecordNotDeletedException("movied id - " + movieId + " not found");
 		} catch (DataAccessException ex) {
-			throw new RecordNotDeletedException("Failed to Delete", ex.getCause());
+			throw new ServiceException("Failed to Delete id-" + movieId, ex.getCause());
 		}
 
 	}
 
 	@Override
-	public String updateMovie(Movie movie) throws RecordNotUpdatedException {
+	public Movie updateMovie(Movie movie) throws RecordNotUpdatedException, ServiceException {
 		try {
+			if (getMovieById(movie.getId()) != null) {
+				movie.getGenre().forEach(val -> {
+					Optional<Integer> genreId = genreRepository.getIdByName(val.getName());
+					if (genreId.isPresent()) {
+						val.setId(genreId.get());
+					}
+				});
+				movie.getGenre().forEach(v -> genreRepository.save(v));
 
-			movie.getGenre().forEach(val->{
-				Optional<Integer>genreId = genreRepository.getIdByName(val.getName());
-				if(genreId.isPresent()) {
-					val.setId(genreId.get());
-				}
-			});
-			movie.getGenre().forEach(v -> genreRepository.save(v));
+				List<Cast> cast = movie.getCast();
+				cast.forEach(v -> castRepository.save(v));
 
-			List<Cast> cast = movie.getCast();
-			cast.forEach(v -> castRepository.save(v));
-
-			movieRepository.save(movie);
-			message = "Successfully updated movie of name - " + movie.getName();
-			return message;
+				return movieRepository.save(movie);
+			} else {
+				throw new RecordNotUpdatedException("invalid movie id. cannot update");
+			}
+		} catch (RecordNotFoundException ex) {
+			throw new RecordNotUpdatedException("invalid movie id. cannot update");
 		}
 
 		catch (DataAccessException ex) {
-			throw new RecordNotUpdatedException("Failed to Update", ex.getCause());
+			throw new ServiceException("Failed to Update", ex.getCause());
 		}
 
 	}
 
 	@Override
-	public Movie getMovieById(int movieId) throws RecordNotFoundException {
+	public Movie getMovieById(int movieId) throws RecordNotFoundException, ServiceException {
 
 		try {
 			Optional<Movie> movie = movieRepository.findById(movieId);
 			if (movie.isPresent()) {
 				return movie.get();
 			} else {
-				throw new RecordNotFoundException("No Record to Fetch");
+				throw new RecordNotFoundException("Movie of id - " + movieId + "is not present");
 			}
 		} catch (DataAccessException ex) {
-			throw new RecordNotFoundException("Failed to Fetch", ex.getCause());
+			throw new ServiceException("Failed to Fetch", ex.getCause());
 		}
 	}
 
 	@Override
-	public List<Movie> getAllMovies() throws EmptyListException {
+	public List<Movie> getAllMovies() throws EmptyListException, ServiceException {
 		try {
 			List<Movie> movies = movieRepository.findAll();
-			if (movies.size()>0) {
+			if (movies.size() > 0) {
 				return movies;
 			} else {
 				throw new EmptyListException("No Record to Fetch");
 			}
 		} catch (DataAccessException ex) {
-			throw new EmptyListException("Failed to Fetch", ex.getCause());
+			throw new ServiceException("Failed to Fetch", ex.getCause());
 		}
 	}
